test(core): move broken and unused tests to quarantine
