+++ packages/core/test/core/quarantine/test_config_plugin_package.js
@@ -5,7 +5,7 @@
 'use strict';
 
 const { test } = require('tap');
-const runner = require('../..').runner.runner;
+const runner = require('../../..').runner.runner;
 const path = require('path');
 
 test('Plugin package name inside plugin config', function (t) {
